pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.26.0",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
]
